Refactor: fix some error-prone issues

+++ replacer-backend/replacer-web/src/main/java/es/bvalero/replacer/user/WebUtils.java
@@ -54,7 +54,7 @@ public class WebUtils {
         // Domain: default
         // SameSite is Lax by default, but it fails in some old browsers, so we set it explicitly.
         return ResponseCookie.from(ACCESS_TOKEN_COOKIE, accessToken.toCookieValue())
-            .maxAge((long) 400 * 24 * 3600)
+            .maxAge(400L * 24 * 3600)
             .path("/api")
             .secure(true)
             .httpOnly(true)
